batch: Extract discard reversal
The batch merge module delegates baseline correspondence construction
during discard. It still owns the pass that reverses presence constraints
over realized entries.

That keeps discard-only restoration logic beside merge placement and
candidate enumeration, even though the pass only needs realized entries and
baseline correspondence.

This commit moves reverse presence-constraint restoration into
batch.discard_reversal and imports it through a private merge-local alias.

The next commit will retarget the discard-reversal coverage to the new
module.
